bouton delete account

+++ frontend/react-projet7/src/pages/Account.js
@@ -13,17 +13,21 @@ const Account = () => {
   };
 
   const deleteUser = (e) => {
-    fetch("http://localhost:8080/api/auth/delete/", {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Bearer ${dataToken}`,
-      },
-      body: JSON.stringify(e),
-    }).then(() => this.setState({ status: "Delete successful" }));
     e.preventDefault();
-    localStorage.clear();
-    window.location = "../connexion";
+    if (
+      window.confirm("Êtes vous sure de vouloir supprimer votre compte ?")
+    ) {
+      fetch("http://localhost:8080/api/auth/delete/", {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: `Bearer ${dataToken}`,
+        },
+      }).then(() => {
+        localStorage.clear();
+        window.location = "../inscription";
+      });
+    }
   };
 
   return (
@@ -37,24 +41,18 @@ const Account = () => {
           className="button-form-account"
           type="submit"
           onClick={buttonLogout}
-        >Déconnexion<i class="fas fa-sign-out-alt"></i>
+        >
+          Déconnexion<i class="fas fa-sign-out-alt"></i>
         </button>
       </div>
       <PostToFill content="" />
-      <form>
-        <button className="button-delete-account"
-          onClick={() => {
-            if (
-              window.confirm(
-                "Êtes vous sure de vouloir supprimer votre compte ?"
-              )
-            )
-              deleteUser();
-          }}
-        >
-          Supprimer compte <i class="fas fa-user-slash"></i>
-        </button>
-      </form>
+      <button
+        className="button-delete-account"
+        type="button"
+        onClick={deleteUser}
+      >
+        Supprimer compte <i class="fas fa-user-slash"></i>
+      </button>
     </div>
   );
 };
